+++ source/mathPres/MathCAT/preferences.py
@@ -40,18 +40,17 @@ class DecimalSeparatorOption(DisplayStringStrEnum):
 	DOT = "."
 	COMMA = ","
 	CUSTOM = "Custom"
+	"""Custom is currently not functional or surfaced in the UI, but is included here for future use."""
 
 	@property
 	def _displayStringLabels(self) -> dict["DecimalSeparatorOption", str]:
 		return {
 			# Translators: options for decimal separator -- "Auto" = automatically pick the choice based on the language
 			self.AUTO: pgettext("math", "Automatic"),
-			# options for decimal separator -- use "."  (and use ", " for block separators)
 			self.DOT: ".",
-			# options for decimal separator -- use ","  (and use ". " for block separators)
 			self.COMMA: ",",
 			# Translators: options for decimal separator -- "Custom" = user sets it
-			#   Currently there is no UI for how it is done yet, but eventually there will be a dialog that pops up to set it
+			# Custom is currently not functional or surfaced in the UI, but is included here for future use.
 			self.CUSTOM: pgettext("math", "Custom"),
 		}